test: extract util
